requirements.txt
configparser
requests
urllib
app_store_scraper 
google_play_scraper 
pandas
logging
jira 
supabase